git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/branches/0.3.0@1029 72d1a826-0e72-11de-b5bf-434079a3519f

+++ mst-common/src/java/xc/mst/repo/sql/create_harvest_repo.sql
@@ -1,5 +1,5 @@
 create table repo_name.record_oai_ids (
-    record_id        int          NOT NULL,
+    record_id        BIGINT       NOT NULL,
     oai_id           varchar(255) not null,
     
 	PRIMARY KEY (record_id)
